Issue-2 refactor: Entity class package 위치 변경
