+++ app/src/main/java/com/drunkenboys/calendarun/data/checkpoint/entity/CheckPoint.kt
@@ -1,12 +1,13 @@
-package com.drunkenboys.calendarun.data.local.calendar.entity
+package com.drunkenboys.calendarun.data.checkpoint.entity
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
+import com.drunkenboys.calendarun.data.calendar.entity.Calendar
 import java.util.*
 
 @Entity(
-    tableName = "Schedule",
+    tableName = "checkpoint",
     foreignKeys = [
         ForeignKey(
             entity = Calendar::class,
@@ -16,14 +17,12 @@ import java.util.*
         )
     ]
 )
-data class Schedule(
+data class CheckPoint(
     @PrimaryKey val id: Int,
     val calendarId: Int,
     val name: String,
     val startDate: Date,
     val endDate: Date,
-    val notification: Date,
-    val memo: String,
-    // TODO <Library>.ScheduleColorType
     val color: Int
 )
+
