+++ app/src/main/java/com/drunkenboys/calendarun/data/schedule/entity/Schedule.kt
@@ -1,12 +1,13 @@
-package com.drunkenboys.calendarun.data.local.calendar.entity
+package com.drunkenboys.calendarun.data.schedule.entity
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
+import com.drunkenboys.calendarun.data.calendar.entity.Calendar
 import java.util.*
 
 @Entity(
-    tableName = "checkpoint",
+    tableName = "Schedule",
     foreignKeys = [
         ForeignKey(
             entity = Calendar::class,
@@ -16,12 +17,14 @@ import java.util.*
         )
     ]
 )
-data class CheckPoint(
+data class Schedule(
     @PrimaryKey val id: Int,
     val calendarId: Int,
     val name: String,
     val startDate: Date,
     val endDate: Date,
+    val notification: Date,
+    val memo: String,
+    // TODO <Library>.ScheduleColorType
     val color: Int
 )
-
